# reference/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bottega API",
    "version": "1.12.0",
    "description": "REST + WebSocket API for Bottega — a task-driven interface to the Claude Code CLI and other coding agents.\n\nThis spec is **hand-maintained** (the routes don't yet share a central registry). When you add or change a route, update this file. It is served at `/api/openapi.json` and rendered at `/api/docs` in both normal and headless modes.\n\n**Auth:** all endpoints require a bearer token (`Authorization: Bearer <jwt|ccui_...>`) except `/health`, `/api/auth/status`, `/api/auth/register`, `/api/auth/login`, `/api/app-settings`, `/api/webhooks/github`, and the docs/spec endpoints. Bootstrap a fresh instance with `GET /api/auth/status` -> `POST /api/auth/register` -> use the returned `token`.\n\n**WebSocket:** connect to `/ws?token=<jwt>` for conversation streaming and live task/conversation subscriptions (not described by this OpenAPI document)."
  },
  "servers": [{ "url": "/", "description": "This server" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/projects": {
      "get": { "tags": ["projects"], "summary": "List projects", "responses": { "200": { "description": "Projects", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } } } } },
      "post": { "tags": ["projects"], "summary": "Create a project", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "repoFolderPath": { "type": "string" }, "subprojectPath": { "type": "string" } }, "required": ["name", "repoFolderPath"] } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } } } }
    },
...
    "/api/projects/{id}/files": { "get": { "tags": ["projects"], "summary": "Project file tree", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
...
    "/api/transcribe": { "post": { "tags": ["conversations"], "summary": "Transcribe audio (requires OPENAI_API_KEY)", "requestBody": { "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "audio": { "type": "string", "format": "binary" } } } } } }, "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
...
    "/api/app-settings": {
      "get": { "tags": ["settings"], "summary": "Get global app settings", "security": [], "responses": { "200": { "$ref": "#/components/responses/Json" } } },
      "put": { "tags": ["settings"], "summary": "Update global app settings", "security": [], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "internal_tool_name": { "type": "string" }, "github_pr_trigger": { "type": "string" } } } } } }, "responses": { "200": { "$ref": "#/components/responses/Json" }, "400": { "$ref": "#/components/responses/Error" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "App JWT or per-user API key (`ccui_...`). Also accepted as `?token=` for WebSocket / media URLs." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
    },
    "responses": {
      "Json": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } },
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
    },
    "schemas": {
      "ApiError": { "type": "object", "properties": { "error": { "type": "string" } }, "required": ["error"] },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" }, "user_id": { "type": "integer" }, "name": { "type": "string" },
          "repo_folder_path": { "type": "string" }, "subproject_path": { "type": ["string", "null"] },
          "active_worktree_task_id": { "type": ["integer", "null"] }, "serve_symlink_path": { "type": ["string", "null"] },
          "systemd_service_name": { "type": ["string", "null"] }, "app_url": { "type": ["string", "null"] },
          "created_at": { "type": "string" }, "updated_at": { "type": "string" }
        }
      },
...
    }
  }
}